Add custom command and desktop margin adaption
Add a DONE_COMMAND array variable for a custom command, executed when the countdown reaches zero. This can be used for playing sounds or showing notifications, etc.

Add a `--margin MARGIN` option to enable automatic adaptation of the desktop margin on the desktop, the program is started on. It works only in XFCE though.